+++ CHANGELOG.md
@@ -4,10 +4,13 @@
 
 ### Added
 
+- concord-tasks: support for file attachments when starting new
+processes;
 - concord-sdk, crypto-tasks: expose `encryptString` method to flows;
 - slack: support for creating and replying to threads;
 - concord-agent, concord-runner: support for a configurable list of
 volumes to mount into Docker containers created by plugins.
+- concord-tasks: support for file attachments when starting a process
 
 ### Changed